fix: 칵테일 좋아요 요청 Request 수정

+++ backend/src/main/java/com/ssafy/cocktail/backend/cocktails/dto/request/CocktailIDReq.java
@@ -6,5 +6,5 @@ import io.swagger.v3.oas.annotations.media.Schema;
 public class CocktailIDReq {
     @Schema(defaultValue = "칵테일 ID", example = "0")
     @JsonProperty("cocktail_id")
-    private Double cocktailId;
+    private Long cocktailId;
 }
